SKERN-1765: NMD-service: install via Helm
This task is to convert the NMD-service to install via Helm. The proposed approach is to change only the nmd-service installation to Helm and continue using Ansible to set up the interim NFS solution until the new storage approach is implemented by SKERN-1881.

Test procedure: Verify that nmd-service installs via Helm