Snyk does not auto select `setup.py`
This also installs `manifestgen` before invoking Snyk as to ensure all
dependencies are present for the scan.

Snyk also pulls `setup.py` into a temporary directory with no other
files present. In doing so, it can't resolve either the `.git` nor
`PKG-INFO` directory as it reads `setup.py`. Snyk was failing to scan
`manifestgen` with the recent build on tag merge (#15), to fix this the
`get_version()` function now has a case to check the installed
`manifestgen` with `pkg_resources`.